pipeline/news-pipeline/consumer/requirements.txt
psycopg2-binary==2.9.10
python-dateutil==2.9.0
pydantic==1.10.15
openai==1.61.1
tiktoken==0.8.0
numpy==1.23.5
python-dotenv==1.0.1
kafka-python
elasticsearch==8.17.1
loguru
